Fix refcounts for MountSources.
fs.NewInode was incrementing the MountSource refcount, but there are lots of
places where we pass a MountSource+ref to NewInode. I considered changing those
call sites to decrement the MountSource referenc, but it seemed easier to
change NewInode such that it inherits a reference. That is also more aligned
with how NewDirent and NewFile work.

PiperOrigin-RevId: 261189211